+++ app/(tabs)/Storage.tsx
@@ -1,23 +1,11 @@
 import { CustomText } from '@/components/CustomText';
 import { Icon } from '@/components/icons';
 import { SafeScreenLayout } from '@/components/layout/SafeScreenLayout';
+import EmotionCardList from '@/components/storage/emotionCardList';
 import Filter from '@/components/storage/Filter';
 import { GreyColors } from '@/constants/Colors';
-import { getGraphicUrl } from '@/constants/graphic';
 import React from 'react';
-import {
-  FlatList, // ✅ 추가
-  ImageBackground,
-  Pressable,
-  StyleSheet,
-  View,
-} from 'react-native';
-
-const DATA = Array.from({ length: 12 }, (_, i) => ({
-  id: String(i),
-  date: '25.08.07',
-  kind: 'awkward' as const,
-}));
+import { StyleSheet, View } from 'react-native';
 
 export default function Storage() {
   return (
@@ -46,38 +34,7 @@ export default function Storage() {
         </View>
       </View>
 
-      <FlatList
-        data={DATA}
-        keyExtractor={(item) => item.id}
-        numColumns={2}
-        showsVerticalScrollIndicator={false}
-        contentContainerStyle={styles.gridContent}
-        columnWrapperStyle={{ gap: 16 }}
-        renderItem={({ item }) => (
-          <Pressable
-            onPress={() => {
-              /** TODO : 마음쪽지 페이지 이동 */
-            }}
-            style={styles.cardContainer}
-          >
-            <ImageBackground
-              source={{
-                uri: getGraphicUrl({ kind: item.kind, page: 'storage' }),
-              }}
-              style={styles.card}
-              imageStyle={styles.imageRadius}
-            >
-              <CustomText
-                variant='body3'
-                color={GreyColors.grey500}
-                style={styles.date}
-              >
-                {item.date}
-              </CustomText>
-            </ImageBackground>
-          </Pressable>
-        )}
-      />
+      <EmotionCardList />
     </SafeScreenLayout>
   );
 }
@@ -99,32 +56,4 @@ const styles = StyleSheet.create({
     justifyContent: 'center',
     alignItems: 'center',
   },
-
-  gridContent: {
-    paddingBottom: 24,
-    gap: 16,
-  },
-  cardContainer: {
-    flex: 1,
-    aspectRatio: 1,
-    borderRadius: 16,
-    shadowColor: '#000',
-    shadowOffset: { width: 0, height: 2 },
-    shadowOpacity: 0.05,
-    shadowRadius: 20,
-    elevation: 4,
-
-    borderWidth: 1,
-    borderColor: '#ffffff',
-  },
-
-  card: {
-    flex: 1,
-    borderRadius: 16,
-    overflow: 'hidden', // 여기선 이미지/내용만 잘라줌
-    justifyContent: 'flex-end',
-  },
-  imageRadius: { borderRadius: 12 },
-
-  date: { paddingVertical: 10, paddingHorizontal: 12 },
 });
