+++ components/storage/EmotionCardList.tsx
@@ -1,0 +1,78 @@
+import { GreyColors } from '@/constants/Colors';
+import { getGraphicUrl } from '@/constants/graphic';
+import { ImageBackground, Pressable, StyleSheet } from 'react-native';
+import { CustomText } from '../CustomText';
+import GridList from '../ui/gridList';
+import { useSavedNotesQuery } from './module/useSavedNotesQuery';
+
+const EmotionCardList = () => {
+  const { data, isLoading, isError } = useSavedNotesQuery();
+
+  if (isLoading) {
+    return null;
+  }
+  if (isError || !data) {
+    return null;
+  }
+
+  return (
+    <GridList
+      data={data}
+      numColumns={2}
+      gap={16}
+      keyExtractor={(item) => item.id}
+      contentContainerStyle={styles.gridContent}
+      renderItem={({ item }) => (
+        <Pressable style={styles.cardContainer}>
+          <ImageBackground
+            source={{
+              uri: getGraphicUrl({ kind: item.kind, page: 'storage' }),
+            }}
+            style={styles.card}
+            imageStyle={styles.imageRadius}
+          >
+            <CustomText
+              variant='body3'
+              color={GreyColors.grey500}
+              style={styles.date}
+            >
+              {item.date}
+            </CustomText>
+          </ImageBackground>
+        </Pressable>
+      )}
+    />
+  );
+};
+
+export default EmotionCardList;
+
+const styles = StyleSheet.create({
+  gridContent: {
+    paddingBottom: 24,
+    gap: 16,
+  },
+  cardContainer: {
+    flex: 1,
+    aspectRatio: 1,
+    borderRadius: 16,
+    shadowColor: '#000',
+    shadowOffset: { width: 0, height: 2 },
+    shadowOpacity: 0.05,
+    shadowRadius: 20,
+    elevation: 4,
+
+    borderWidth: 1,
+    borderColor: '#ffffff',
+  },
+
+  card: {
+    flex: 1,
+    borderRadius: 16,
+    overflow: 'hidden', // 여기선 이미지/내용만 잘라줌
+    justifyContent: 'flex-end',
+  },
+  imageRadius: { borderRadius: 12 },
+
+  date: { paddingVertical: 10, paddingHorizontal: 12 },
+});
